+++ shared_lib_tests/shared_lib_sssp.c
@@ -24,7 +24,7 @@ int main(int argc, char* argv[]) {
     config.traversal_mode     =     0;     // 0 for Load balanced partition
     config.max_queue_sizing   =  1.0f;     // maximum queue sizing factor
 
-    size_t num_nodes = 7, num_edges = 15;  // number of nodes and edges
+    int num_nodes = 7, num_edges = 15;  // number of nodes and edges
     int row_offsets[8]  = {0, 3, 6, 9, 11, 14, 15, 15};
     int col_indices[15] = {1, 2, 3, 0, 2, 4, 3, 4, 5, 5, 6, 2, 5, 6, 6};
     int edge_values[15] = {39, 6, 41, 51, 63, 17, 10, 44, 41, 13, 58, 43, 50, 59, 35};
@@ -42,7 +42,7 @@ int main(int argc, char* argv[]) {
     ////////////////////////////////////////////////////////////////////////////
     int *labels = (int*)malloc(sizeof(int) * graph_i->num_nodes);
     labels = (int*)graph_o->node_value1;
-    size_t node; for (node = 0; node < graph_i->num_nodes; ++node)
+    int node; for (node = 0; node < graph_i->num_nodes; ++node)
         printf("Node_ID [%d] : Label [%d]\n", node, labels[node]);
 
     if (graph_i) free(graph_i);
